// sbsp_frontend/src/types/GlobalHostSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppearanceSettings } from './AppearanceSettings';
import type { AudioHardwareSettings } from './AudioHardwareSettings';
import type { GeneralSettings } from './GeneralSettings';
import type { HotkeySettings } from './HotkeySettings';
import type { NameFormatSettings } from './NameFormatSettings';
import type { TemplateSettings } from './TemplateSettings';

export type GlobalHostSettings = {
  general: GeneralSettings;
  audio: AudioHardwareSettings;
  appearance: AppearanceSettings;
  hotkey: HotkeySettings;
  template: TemplateSettings;
  nameFormat: NameFormatSettings;
};