// sbsp_frontend/src/types/FullShowState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProjectStatus } from "./ProjectStatus";
import type { ShowModel } from "./ShowModel";
import type { ShowState } from "./ShowState";

export type FullShowState = { projectStatus: ProjectStatus, showModel: ShowModel, showState: ShowState, };