// sbsp_frontend/src/types/BackendError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BackendError = { "type": "fileSave", path: string, message: string, } | { "type": "fileLoad", path: string, message: string, } | { "type": "cueEdit", message: string, } | { "type": "custom", id: number, message: string, };