// sbsp_frontend/src/types/InsertPosition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type InsertPosition = { "type": "before", target: string, } | { "type": "after", target: string, } | { "type": "inside", target: string | null, index: number, } | { "type": "last" };